Change table and file name of Talks
* In order to improve readability and consistency towards
service. The old name assistants and attends(file) are changed
to assists and assist(file). This because in that table we are creating
a relation between two user and a talk given by one of those two users.
This relation is an assitance to a talk from the other one user left.

* Also, change the column names speaker and assitand and add to them
a sufix to specify that they are an id